change type

+++ src/Utilities.jl
@@ -21,7 +21,7 @@ function indices(n_a, n_h, n_zl, n_zh, n_zm,ind)
 end
 
 const to = TimerOutput()
-function prettytime(t)
+function prettytime(t::Int)
     if t < 1e3
         value, units = t, "ns"
     elseif t < 1e6
